Check if name is set, do not assert
As the `name` parameter is passed to the method from the outside (i.e.
it could contained anything passed to the library from the script using
it), it does not make sense to `assert` it.

Using `assert` also has the drawback that it will be a no-op in
production environments (i.e. when `__debug__` is `False`), so the
method would generate XML which would violate the protocol spec and
would be refused by the server.

Instead, the method now checks if `name` is `True`, e.g. a non-empty
string, and raises a ValueError if not.