Remove commented-out debug statements
This commit removes two instances of statements which were commented out
but preceded by a comment indicating they should be used for debugging.

As a general rule, if additional logging should be needed in GMP scripts
it should be controlled through parameters passed to `gvm_pyshell` and
not require source code changes.