Allow to transform command responses
Enhance Gmp class to pass a transform callable to init. The transform
callable is will be called for each response and can convert or check
the response.

Currently there are three kinds of transform callable to cover the old
use cases of shell mode, raw mode and raising an exception on responses
containing an error status.