Use equality tests instead of object identity tests
This commit addresses a number of instances in the GMP scripts where the
`is` and `is not` operators were used in comparisons.

In all cases, a comparison based on the value of the objects was
desired. However, `is` and `is not` test for object identity and need to
be replaced with the `==` and `!=` to test for object value.