Improve config file format and settings loading
The config file may look like the following now:

[main]
timeout=120

[gmp]
username=foo
password=bar

[ssh]
username=foo
password=bar
port=1234

[tls]
certfile=/path/to/certfile
cafile=/path/to/cafile
keyfile=/path/to/keyfile
port=1234

[unixsocket]
socketpath=/path/to/unix/socket

Interpolation and fallback to main section as described in
https://docs.python.org/3.5/library/configparser.html is also possible.

Example to use the same username for ssh and gmp:
[main]
username=foo

[gmp]

[ssh]

For this use case it is important that the sections for gmp and ssh are
available. If not present the configparser won't fallback to the main
section.