Add monthly-report ++reports option, fix report id
The monthly-report script now looks for reports in identifiers
modified in the given month.
A new ++reports option has been added to choose whether to skip
getting the reports (only returning the vulnerability counts), get the
last report in the selected month or get a list of all reports in the
selected month.
For this the argument parsing and help/usage text are changed to use
argparse.

This addresses the problem that the shown report ids could previously
be outside the selected date range.